Rewrite to be run at logon, before Explorer opens
This variant doesn't kill Explorer, instead it can be run at logon in the background. When Explorer launches it relinquishes the hotkeys for other programs to use and tries to take them back if Explorer exits for any reason. Either because Explorer crashes or due to a logoff event.

See also:
 - https://stackoverflow.com/a/24646860
 - https://gist.github.com/rdp/6b5fc8993089ee12b44d
 - https://stackoverflow.com/a/31020562
 - http://www.cplusplus.com/forum/windows/12137/